+++ Yahtzee/Y_NN.py
@@ -284,7 +284,7 @@ class MCTSParallel:
     @torch.no_grad()
     def search(self,states,player,spGames,last_action=None,throw=0):
         #dirichlet variant
-        _, policy = self.model(torch.tensor(self.game.get_encoded_states(states,throw),device=self.model.device))
+        _, policy = self.model(torch.tensor(self.game.get_encoded_states(states,throw),device=self.model.device,dtype=torch.float32))
         policy = torch.softmax(policy,1).detach().cpu().numpy()
         policy = (1 - self.args['dirichlet_epsilon']) * policy + self.args['dirichlet_epsilon'] * np.random.dirichlet([self.args['dirichlet_alpha']] * len(policy[0]), size=policy.shape[0])
         for i,spg in enumerate(spGames):
